Use Laravel query builder for DELETE statements when possible (#1729)
We currently have several pre-laravel methods which can execute DELETE
queries, but it is preferable to use Laravel's database functions
instead. One advantage is that all queries executed by laravel are
visible in Laravel's query log, which helps debug performance issues.
DELETE statements are an easy starting place because the return value is
rarely used.

+++ app/cdash/app/Model/User.php
@@ -16,6 +16,7 @@
 namespace CDash\Model;
 
 use CDash\Database;
+use Illuminate\Support\Facades\DB;
 
 class User
 {
@@ -169,8 +170,7 @@ class User
         if (!$this->Id) {
             return false;
         }
-        $stmt = $this->PDO->prepare("DELETE FROM $this->TableName WHERE id = ?");
-        pdo_execute($stmt, [$this->Id]);
+        DB::delete("DELETE FROM $this->TableName WHERE id = ?", [$this->Id]);
     }
 
     /** Get the email */
@@ -314,12 +314,7 @@ class User
                 $row = $stmt->fetch();
                 $cutoff = $row['date'];
                 // Then delete the ones that are too old
-                $stmt = $this->PDO->prepare(
-                    'DELETE FROM password
-                    WHERE userid= :userid AND date < :date');
-                $stmt->bindParam(':userid', $this->Id);
-                $stmt->bindParam(':date', $cutoff);
-                pdo_execute($stmt);
+                DB::delete('DELETE FROM password WHERE userid = ? AND date < ?', [$this->Id, $cutoff]);
             }
         }
     }
